+++ ppmd/coulomb/FMMSource/TranslateLTL.cpp
@@ -1,7 +1,7 @@
 
 
 static inline void lin_to_xyz(
-    const UINT32 * RESTRICT dim_parent,
+    const INT64 * RESTRICT dim_parent,
     const INT64 lin,
     INT64 * RESTRICT cx,
     INT64 * RESTRICT cy,
@@ -14,7 +14,7 @@ static inline void lin_to_xyz(
 }
 
 static inline INT64 xyz_to_lin(
-    const UINT32 * RESTRICT dim_child,
+    const INT64 * RESTRICT dim_child,
     const INT64 cx,
     const INT64 cy,
     const INT64 cz
@@ -184,8 +184,8 @@ static inline void ltl_octal(
 
 extern "C"
 int translate_ltl(
-    const UINT32 * RESTRICT dim_parent,     // slowest to fastest
-    const UINT32 * RESTRICT dim_child,      // slowest to fastest
+    const INT64 * RESTRICT dim_parent,     // slowest to fastest
+    const INT64 * RESTRICT dim_child,      // slowest to fastest
     REAL * RESTRICT moments_child,
     const REAL * RESTRICT moments_parent,
     const REAL * RESTRICT ylm,
