+++ ppmd/coulomb/FMMSource/TranslateMTLZ2.cpp
@@ -311,7 +311,7 @@ int wrapper_rotate_moments_backward(
 
 
 static inline void lin_to_xyz(
-    const UINT32 * RESTRICT dim_parent,
+    const INT64 * RESTRICT dim_parent,
     const INT64 lin,
     INT64 * RESTRICT cx,
     INT64 * RESTRICT cy,
@@ -324,7 +324,7 @@ static inline void lin_to_xyz(
 }
 
 static inline INT64 xyz_to_lin(
-    const UINT32 * RESTRICT dim_child,
+    const INT64 * RESTRICT dim_child,
     const INT64 cx,
     const INT64 cy,
     const INT64 cz
@@ -947,7 +947,7 @@ int wrapper_blocked_mtl_z(
 
 extern "C"
 int translate_mtl(
-    const UINT32 * RESTRICT dim_child,      // slowest to fastest
+    const INT64 * RESTRICT dim_child,      // slowest to fastest
     const REAL * RESTRICT multipole_moments,
     REAL * RESTRICT local_moments,
     const REAL * RESTRICT const * RESTRICT const * RESTRICT wig_forw,
@@ -972,9 +972,9 @@ int translate_mtl(
     const INT64 ncomp2 = nlevel*nlevel*8;
     const INT64 im_offset = nlevel*nlevel;
     const INT64 im_offset2 = 4*nlevel*nlevel;
-    const UINT32 dim_halo[3] = {dim_child[0] + 4,
+    const INT64 dim_halo[3] = {dim_child[0] + 4,
         dim_child[1] + 4, dim_child[2] + 4};
-    const UINT32 dim_eight[3] = {2, 2, 2};
+    const INT64 dim_eight[3] = {2, 2, 2};
 
     const INT32 phi_stride = 8*nlevel + 2;
     const INT32 theta_stride = 4 * nlevel * nlevel;
